feat: add u220

+++ src/openzeppelin/utils/structs/checkpoints.cairo
@@ -25,13 +25,13 @@ impl TraceImpl of TraceTrait {
     /// Pushes a (`key`, `value`) pair into a Trace so that it is stored as the checkpoint.
     ///
     /// Returns previous value and new value.
-    fn push(ref self: Trace, key: u64, value: u220) -> (u220, u220) {
+    fn push(ref self: Trace, key: u32, value: u220) -> (u220, u220) {
         self._checkpoints._insert(key, value)
     }
 
     /// Returns the value in the first (oldest) checkpoint with key greater or equal
     /// than the search key, or zero if there is none.
-    fn lower_lookup(self: Trace, key: u64) -> u220 {
+    fn lower_lookup(self: Trace, key: u32) -> u220 {
         let mut checkpoints = self._checkpoints;
         let len = checkpoints.len();
         let pos = checkpoints._lower_binary_lookup(key, 0, len);
@@ -47,7 +47,7 @@ impl TraceImpl of TraceTrait {
 
     /// Returns the value in the last (most recent) checkpoint with key lower or equal
     /// than the search key, or zero if there is none.
-    fn upper_lookup(self: Trace, key: u64) -> u220 {
+    fn upper_lookup(self: Trace, key: u32) -> u220 {
         let mut checkpoints = self._checkpoints;
         let len = checkpoints.len();
         let pos = checkpoints._upper_binary_lookup(key, 0, len);
@@ -66,7 +66,7 @@ impl TraceImpl of TraceTrait {
     ///
     /// NOTE: This is a variant of {upper_lookup} that is optimised to
     /// find "recent" checkpoint (checkpoints with high keys).
-    fn upper_lookup_recent(self: Trace, key: u64) -> u220 {
+    fn upper_lookup_recent(self: Trace, key: u32) -> u220 {
         let mut checkpoints = self._checkpoints;
         let len = checkpoints.len();
 
@@ -111,7 +111,7 @@ impl TraceImpl of TraceTrait {
 
     /// Returns whether there is a checkpoint in the structure (i.e. it is not empty),
     /// and if so the key and value in the most recent checkpoint.
-    fn latest_checkpoint(self: Trace) -> (bool, u64, u220) {
+    fn latest_checkpoint(self: Trace) -> (bool, u32, u220) {
         let mut checkpoints = self._checkpoints;
         let pos = checkpoints.len();
 
@@ -140,7 +140,7 @@ impl TraceImpl of TraceTrait {
 impl CheckpointImpl of CheckpointTrait {
     /// Pushes a (`key`, `value`) pair into an ordered list of checkpoints, either by inserting a new checkpoint,
     /// or by updating the last one.
-    fn _insert(ref self: Vec<Checkpoint>, key: u64, value: u220) -> (u220, u220) {
+    fn _insert(ref self: Vec<Checkpoint>, key: u32, value: u220) -> (u220, u220) {
         let pos = self.len();
 
         if (pos > 0) {
@@ -167,7 +167,7 @@ impl CheckpointImpl of CheckpointTrait {
     /// Return the index of the last (most recent) checkpoint with key lower or equal than the search key,
     /// or `high` if there is none. `low` and `high` define a section where to do the search, with
     /// inclusive `low` and exclusive `high`.
-    fn _upper_binary_lookup(ref self: Vec<Checkpoint>, key: u64, low: u32, high: u32) -> u32 {
+    fn _upper_binary_lookup(ref self: Vec<Checkpoint>, key: u32, low: u32, high: u32) -> u32 {
         let mut _low = low;
         let mut _high = high;
         loop {
@@ -189,7 +189,7 @@ impl CheckpointImpl of CheckpointTrait {
     /// Return the index of the first (oldest) checkpoint with key is greater or equal than the search key,
     /// or `high` if there is none. `low` and `high` define a section where to do the search, with
     /// inclusive `low` and exclusive `high`.
-    fn _lower_binary_lookup(ref self: Vec<Checkpoint>, key: u64, low: u32, high: u32) -> u32 {
+    fn _lower_binary_lookup(ref self: Vec<Checkpoint>, key: u32, low: u32, high: u32) -> u32 {
         let mut _low = low;
         let mut _high = high;
         loop {
